rename to client

+++ src/main/java/com/etienne/client/store/model/Eye.java
@@ -11,7 +11,7 @@ public class Eye {
 
     private Double dioptria;
     private Double cilinder;
-    private Integer fok;
+    private Double fok;
     private Double vizus;
 
 }
